backend/requirements.txt
fastapi
uvicorn[standard] 

# Supabase client
supabase
python-dotenv
pytest
requests

# Authentication
PyJWT
cryptography
python-multipart

# Testing dependencies  
httpx
pytest-asyncio